handle cancel event
I've been using this class for cells in a uitableview, and there are certain cases when the touch ends up cancelled instead of ended. This causes the button to get stuck a state as if it were permanently pressed down.

Consequently I took the code that handled an ended touch and overrode the cancel handler to perform the same action.